add function validateBody
