+++ models/contacts.json
@@ -19,15 +19,15 @@
   },
   {
     "id": "vza2RIzNGIwutCVCs4mCL",
-    "name": "Wylie Pope",
+    "name": "Wylow Pope",
     "email": "est@utquamvel.net",
     "phone": "(692) 802-2949"
   },
   {
     "id": "05olLMgyVQdWRwgKfg5J6",
-    "name": "Cyrus JacksoN",
+    "name": "Cyru JacksoN",
     "email": "nibh@semsempererat.com",
-    "phone": "(500) 477-5218"
+    "phone": "(500) 47-5218"
   },
   {
     "id": "1DEXoP8AuCGYc1YgoQ6hw",
