+++ routes/api/contacts.js
@@ -2,16 +2,19 @@ const express = require("express");
 
 const ctrl = require("../../controllers/contacts");
 
+const validateBody = require("../../middlewares/validateBody");
+const addSchema = require("../../schemas/contacts");
+
 const router = express.Router();
 
 router.get("/", ctrl.listContacts);
 
 router.get("/:id", ctrl.getContactById);
 
-router.post("/", ctrl.addContact);
+router.post("/", validateBody(addSchema), ctrl.addContact);
 
 router.delete("/:id", ctrl.removeContact);
 
-router.put("/:id", ctrl.updateContact);
+router.put("/:id", validateBody(addSchema), ctrl.updateContact);
 
 module.exports = router;
